docs(custom_reactivity): add a note to clarify the name mapping between Ruby/JS components
Found it very intuitive to understand where we're mapping and after some clarifications from the
core team, understanding how the Ruby component picks up the right Vue JS component, was a real step
forward in understanding the underlying implementation and removes a bit of the black-box magic.
Personally, I find this a very neat fact to know. It might have been mentioned above but tried to
make it as concrete and digestible in a little helpful info-box.