Emit Multiple Events with `onclick` component (#10)
* [ISSUE-5] Add spec coverage for onclick component

Sets up a spec coverage suite for the new multi-event emission
behavior for the onclick component as well as a slight refactor of
the existing specs.

* [ISSUE-5] Add multi-event emission ability to onclick component

Adds the multi-event emission ability to the onclick component, while
abstracting this ability into a mixin that can be used by other
components as well.

* [ISSUE-5] Update documentation for the onclick event's API

There was an outdated section that mentioned that event payload
can't be bundled in with the emitted event. Updated
this as it's part of the API already. Also, covered the multi-event
emission feature that's being added.

The new coverage is on both documentation and specs. Making sure that
the uncovered symbol/string event name is documented properly, examples
provided and an accompanying spec that demonstrates the expected behavior.