[ISSUE-6] Refactor the toggle component
Updates the documentation with some less than clear text. Refactoring
the specs for context clarity and better readability and documentation
capabilities. As a side, refactored the internals of the toggle
component's setup for cleanliness and code reuse of the already provided
functionality in `mixin.js`.

As an attempt to start "prettying-up" the dummy app, added a section
specifically for testing and playing around with the toggle component.