[ISSUE-7] Update transition component's ability to emit a custom event
Adding in the `emit` property as optional and forwarding it inside
`vue_props`. On the flip-side within the VueJs component, using the
`emitEvents` function defined in mixin.js as the standard.

I'm favoring a convention where if we define `const self = this` inside
a function, we use that as the reference to the component throughout
the entire function instead of mixing and matching `self` and `this`
interchangeably in a single method. I find this much clearer and
helps to keep less context on top of mind while trying to grok a
piece of code.