Add slot-ability to `collection_order_toggle_indicator`
Adds the ability to provide slots to the collection order toggle
indicator in order to improve better rendering capabilities as to
what I can set as an indicator instead of just unescaped text.

The default rendering of a key such as `asc` is now plaintext instead
of unescaped text.

A spec was added to demonstrate the unescaped text feature is still
possible by utilizing a slot, so an upgrade will be needed if relying
on the unescaped default in order to preserve your current functionality.